[REFACTOR] 영상 추천결과를 redis에 저장하는 함수 위치 변경
- 기존 redis조회 repository에서 담당
