+++ src/main/java/com/example/whiplash/recommend/youtube/repository/YoutubeRecommendRedisRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Set;
 
 /**
- * 유튜브 추천 영상을 Redis에서 조회하는 Repository
+ * 유튜브 추천 영상을 Redis에서 조회/저장하는 Repository
  */
 @Slf4j
 @RequiredArgsConstructor
@@ -73,6 +73,37 @@ public class YoutubeRecommendRedisRepository {
         } catch (Exception e) {
             log.error("Error while fetching videos from Redis for key: {}", key, e);
             return new ArrayList<>();
+        }
+    }
+
+    /**
+     * 키워드 ID에 대한 추천 영상 목록을 Redis에 저장
+     *
+     * @param keywordId 키워드 ID
+     * @param videos    추천 영상 목록
+     */
+    public void storeRecommendations(Long keywordId, List<YoutubeVideo> videos) {
+        if (videos == null || videos.isEmpty()) {
+            log.warn("No videos to store for keywordId={}", keywordId);
+            return;
+        }
+
+        String key = KEYWORD_RECOMMEND_PREFIX + keywordId;
+
+        try {
+            // 기존 데이터 삭제
+            redisTemplate.delete(key);
+
+            // 새로운 추천 영상 저장 (Set으로 저장)
+            for (YoutubeVideo video : videos) {
+                redisTemplate.opsForSet().add(key, video);
+            }
+
+            log.info("Successfully stored {} videos for keywordId={} in Redis key={}",
+                videos.size(), keywordId, key);
+        } catch (Exception e) {
+            log.error("Failed to store recommendations for keywordId={}", keywordId, e);
+            throw e;
         }
     }
 }
